Add ColorableMergedMaterial class and its test file
